Update with format changes to v4l2-ctl output
My Raspberry PI 4 and cam got some different output that caused it not to run. Older version of v4l-utils maybe? I'm using 1.16.3-3. Modified code so it would work. Output had 'intmenu' in addition to int and menu that caused some mixups. Intmenu is the same as menu I think. Added 'Codec Controls' and 'JPEG Compression Controls'.  Video resolution output was different and included Stepwise 32x32 (which caused the video to be 32x32 when the actual resolution is after that.